# Conflicts:
#	.idea/.gitignore
#	.idea/misc.xml
#	src/main/java/com/system/ElectionManagement/models/Candidate.java
#	src/main/resources/db/data.sql
#	src/test/java/com/system/ElectionManagement/repositories/NominationFormRepositoryTest.java
#	target/classes/application.properties
#	target/classes/com/system/ElectionManagement/config/Config.class
#	target/classes/com/system/ElectionManagement/models/Candidate.class
#	target/classes/com/system/ElectionManagement/models/Election.class
#	target/classes/com/system/ElectionManagement/models/NominationForm.class
#	target/classes/com/system/ElectionManagement/models/Vote.class
#	target/classes/com/system/ElectionManagement/models/Voter.class
#	target/classes/com/system/ElectionManagement/repositories/VoteRepository.class
#	target/classes/db/data.sql
#	target/classes/secrets.properties
#	target/test-classes/com/system/ElectionManagement/repositories/NominationFormRepositoryTest.class

+++ src/main/java/com/system/ElectionManagement/models/ContactInformation.java
@@ -1,9 +1,6 @@
 package com.system.ElectionManagement.models;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToOne;
+import jakarta.persistence.*;
 import lombok.Data;
 
 @Data
@@ -11,7 +8,7 @@ import lombok.Data;
 public class ContactInformation {
     @Id
     @GeneratedValue
-    private int id;
+    private Long id;
     private String phoneNumber;
     private String email;
     @OneToOne
